feat: hallucination reduction — fingerprinting, confidence filtering, changed-file tracking
- Finding.Fingerprint (sha256 of file|title|desc[:50]) embedded in issue bodies
  and used for dedup alongside title fallback
- Finding.Confidence (HIGH/MEDIUM/LOW) parsed from review output; --min-confidence
  flag (default MEDIUM) gates auto-fix and issue filing
- BuildAuditPrompt accepts changedFiles map; iterations 2+ only audit changed files
- RunFix returns (int, []string, error) — exposes changed paths back to loop
- loopState.changedFiles updated after each auto-fix iteration
- ExistingFingerprints() added to github client for cross-run dedup
- MergeMethod/AllowedMergeStrategies/DefaultMergeStrategy moved to types package
- cmd split into run.go / run_bootstrap.go / run_loop.go / run_service_github.go
